fix: products list rendering on home page
